+++ examples/nodejs/client/getEmbeddedSchema.ts
@@ -57,7 +57,9 @@ const client = new ConcordiumGRPCNodeClient(
     const schema = await client.getEmbeddedSchema(moduleRef);
     // #endregion documentation-snippet
 
-    fs.writeFileSync(cli.flags.outPath, schema);
+    if (schema) {
+        fs.writeFileSync(cli.flags.outPath, new Uint8Array(schema.buffer));
+    }
 
     console.log('Wrote schema to file: ', cli.flags.outPath);
 })();