+++ packages/sdk/src/util.ts
@@ -5,6 +5,7 @@ import {
     IpAddressString,
     SchemaVersion,
 } from './types.js';
+import { RawModuleSchema, UnversionedSchemaVersion } from './schemaTypes.js';
 
 /**
  * Replaces a string in a JSON string with the same string as a
@@ -66,65 +67,6 @@ export function countSignatures(
         totalSignatureCount += signatureCount;
     }
     return totalSignatureCount;
-}
-
-export interface SchemaFromWasm {
-    schema: ArrayBuffer;
-    schemaVersion: SchemaVersion | undefined;
-    sectionName: string;
-}
-
-function findCustomSections(m: WebAssembly.Module, moduleVersion: number) {
-    function getCustomSections(
-        sectionName: string,
-        schemaVersion: SchemaVersion | undefined
-    ) {
-        const s = WebAssembly.Module.customSections(m, sectionName);
-        return s.length === 0
-            ? undefined
-            : { sectionName, schemaVersion, contents: s };
-    }
-
-    // First look for section containing schema with embedded version, then "-v1" or "-v2" depending on the module version.
-    switch (moduleVersion) {
-        case 0:
-            return (
-                getCustomSections('concordium-schema', undefined) || // always v0
-                getCustomSections('concordium-schema-v1', SchemaVersion.V0) // v0 (not a typo)
-            );
-        case 1:
-            return (
-                getCustomSections('concordium-schema', undefined) || // v1, v2, or v3
-                getCustomSections('concordium-schema-v2', SchemaVersion.V1) // v1 (not a typo)
-            );
-    }
-    return getCustomSections('concordium-schema', undefined); // expecting to find this section in future module versions
-}
-
-/**
- * Attempts to extract a smart contract schema from a given wasm module, optionally based on the module version.
- * @param moduleSourceWasm the wasm module as a buffer.
- * @param moduleVersion module version. If it's a known version, we will attempt to find the schema in sections specific to that version. Otherwise, we only look in "concordium-schema".
- * @returns the smart contract schema with version and name of custom section where it was found or undefined if no schema was found
- */
-export function wasmToSchema(
-    moduleSourceWasm: BufferSource,
-    moduleVersion: number
-): SchemaFromWasm | undefined {
-    const sections = findCustomSections(
-        new WebAssembly.Module(moduleSourceWasm),
-        moduleVersion
-    );
-    if (!sections) {
-        return undefined;
-    }
-    const { sectionName, schemaVersion, contents } = sections;
-    if (contents.length !== 1) {
-        throw new Error(
-            `invalid module: expected to find at most one custom section named "${sectionName}", but found ${contents.length}`
-        );
-    }
-    return { schema: contents[0], schemaVersion, sectionName };
 }
 
 /**